DH colours and some more scenarios
Need to think the scenarios over a bit more. The one's that are currently there seem a bit lax.

Similarly the font seems kinda off despite checking that the font is the same.

Furthermore, it still looks kinda ugly. I'm not a fan of the round-edged boxes. Think the edges shouldn't be rounded - also the colour is meh